+++ CHANGELOG.md
@@ -21,6 +21,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - **radio**: properties `name` and `mode` and no longer needed and should only be placed on `wcs-radio-group` from now on
+- **editable-field**: rename `errorhandler` to `errorHandler`
 
 ### Deprecated